Move log_event to new file
